structured the code to different directories and split the code into smaller sections (WIP)
