+++ tests/test_extract.py
@@ -1,0 +1,9 @@
+import sys
+import os
+
+current_dir = os.path.dirname(os.path.abspath(__file__))
+sys.path.append(os.path.join(current_dir, '..'))
+
+from src.extract.extract import Extract
+e = Extract(5)
+e.start()
