+++ utils/extract_utils/general.py
@@ -1,0 +1,3 @@
+PAGE_PER_SCAN = 1
+MAX_WORKERS = 10
+MAX_PAGES = 200
