+++ utils/extract_utils/omdb_utils.py
@@ -1,0 +1,3 @@
+from os import getenv
+
+OMDB_API_KEY = getenv('OMDB_API_KEY')
