+++ utils/iterface_utils/kafka_utils.py
@@ -1,0 +1,5 @@
+from os import getenv
+
+CLOUDKARAFKA_HOSTNAME = getenv('CLOUDKARAFKA_HOSTNAME')
+CLOUDKARAFKA_USERNAME = getenv('CLOUDKARAFKA_USERNAME')
+CLOUDKARAFKA_PASSWORD = getenv('CLOUDKARAFKA_PASSWORD')
