--- variables.py
@@ -1,14 +1,0 @@
-from dotenv import load_dotenv
-from os import getenv
-from json import loads
-load_dotenv()
-
-PAGE_PER_SCAN = 1
-MAX_WORKERS = 10
-MAX_PAGES = 200
-OMDB_API_KEY = getenv('OMDB_API_KEY')
-TMDB_HEADERS = loads(getenv('TMDB_HEADERS'))
-
-CLOUDKARAFKA_HOSTNAME = getenv('CLOUDKARAFKA_HOSTNAME')
-CLOUDKARAFKA_USERNAME = getenv('CLOUDKARAFKA_USERNAME')
-CLOUDKARAFKA_PASSWORD = getenv('CLOUDKARAFKA_PASSWORD')
